use both tor and local

+++ money.js
@@ -1,29 +1,31 @@
 const {exec} = require('child_process')
 const puppeteer = require('puppeteer')
 
-const rackupMoney = async (slug, show) => {
+const rackupMoney = async (slug, tor, show) => {
 
-    const browser = await puppeteer.launch(
-        {
-            headless: ! show,
-            args: ['--proxy-server=socks5://127.0.0.1:9050']
+    let launchParameters = {headless: ! show}
+    if (tor) {
+        launchParameters.args = ['--proxy-server=socks5://127.0.0.1:9050']
+    }
+    const browser = await puppeteer.launch(launchParameters)
+
+    if (tor) {
+        const testPage = await browser.newPage()
+        await testPage.goto('https://check.torproject.org/')
+        const isUsingTor = await testPage.$eval('body', el =>
+            el.innerHTML.includes('Congratulations. This browser is configured to use Tor')
+        )
+        
+        if (isUsingTor) {
+            console.log('Puppeteer using tor, all good.')
+        } else {
+            console.log('Not using Tor. Closing...')
+            return await browser.close()
         }
-    )
-
-    const testPage = await browser.newPage()
-    await testPage.goto('https://check.torproject.org/')
-    const isUsingTor = await testPage.$eval('body', el =>
-        el.innerHTML.includes('Congratulations. This browser is configured to use Tor')
-    )
-    
-    if (isUsingTor) {
-        console.log('Puppeteer using tor, all good.')
-    } else {
-        console.log('Not using Tor. Closing...')
-        return await browser.close()
+        
+        testPage.close()
     }
     
-    testPage.close()
 
     try {
         const page = await browser.newPage()
@@ -65,44 +67,59 @@ function letsGo () {
 
     if (process.argv.length < 3) {
 
-        console.warn('Use this program like this: "node money.js <your-utip-slug> <options: show>"')
+        console.warn('Use this program like this: "node money.js <your-utip-slug> <option: tor|both> <option: show>"')
 
     } else {
 
-        console.log('\n\nClear previous tor...\n\n\n\n')
+        const slug = process.argv[2]
+        const tor = process.argv.length > 3 && process.argv[3] === 'tor'
+        const both = process.argv.length > 3 && process.argv[3] === 'both'
+        const show = process.argv.length > 4 && process.argv[4] === 'show'
 
-        exec('taskkill /IM "tor.exe" /F', (error, stdout, stderr) => {
-            if (! error && stdout && ! stderr) {
-                console.log('Tor stopped')
-            } else {
-                console.log('Tor couldn\'t be stopped, probably already stopped')
-            }
+        if (tor || both) {
+            
+            console.log('\n\nClear previous tor...\n\n\n\n')
 
-            console.log('Starting tor...')
-
-            exec('tor&', (error, stdout, stderr) => {
-
+            exec('taskkill /IM "tor.exe" /F', (error, stdout, stderr) => {
                 if (! error && stdout && ! stderr) {
-                    console.log('Tor started')
+                    console.log('Tor stopped')
                 } else {
-                    console.error('Error while starting tor, please make sure tor is started')
+                    console.log('Tor couldn\'t be stopped, probably already stopped')
                 }
+
+                console.log('Starting tor...')
+
+                exec('tor&', (error, stdout, stderr) => {
+
+                    if (! error && stdout && ! stderr) {
+                        console.log('Tor started')
+                    } else {
+                        console.error('Error while starting tor, please make sure tor is started')
+                    }
+
+                })
+
+                setTimeout(() => {
+                    console.log("Let's run an ad on " + slug + "'s page :")
+
+                    rackupMoney(slug, true, show).then(() => {
+                        letsGo()
+                    })
+
+                }, 2000)
 
             })
 
-            setTimeout(() => {
+        }
+        
+        if (! tor || both) {
+            console.log("Let's run an ad on " + slug + "'s page :")
 
-                const slug = process.argv[2]
-                const show = process.argv.length > 3 && process.argv[3] === 'show'
-                console.log("Let's run an ad on " + slug + "'s page :")
-
-                rackupMoney(slug, show).then(() => {
-                    letsGo()
-                })
-
-            }, 2000)
-
-        })
+            rackupMoney(slug, false, show).then(() => {
+                letsGo()
+            })
+        }
+        
     }
 }
